fitness_app/backend/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
Flask-Login==0.5.0
Flask-OAuthlib==0.9.6
psycopg2-binary==2.8.6
requests==2.25.1
boto3==1.17.49
pytest==6.2.2
pytest-cov==2.11.1
coverage==5.5
healthkit==0.2.0
numpy==1.20.2
pandas==1.2.4
python-dotenv==0.15.0
gunicorn==20.0.4
